fixed permission problem for template file
