+++ Controller/FormController.php
@@ -13,9 +13,7 @@ namespace Sulu\Bundle\FormBundle\Controller;
 
 use FOS\RestBundle\Controller\FOSRestController;
 use FOS\RestBundle\Routing\ClassResourceInterface;
-use Sulu\Bundle\FormBundle\Dynamic\FormFieldTypeInterface;
 use Sulu\Bundle\FormBundle\Entity\Form;
-use Sulu\Bundle\FormBundle\Mail\HelperInterface;
 use Sulu\Bundle\FormBundle\Manager\FormManager;
 use Sulu\Component\Rest\ListBuilder\Doctrine\DoctrineListBuilderFactory;
 use Sulu\Component\Rest\ListBuilder\Doctrine\FieldDescriptor\DoctrineCaseFieldDescriptor;
@@ -27,7 +25,6 @@ use Sulu\Component\Rest\RestHelperInterface;
 use Sulu\Component\Security\SecuredControllerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Translation\Translator;
 
 /**
  * Generated by https://github.com/alexander-schranz/sulu-backend-bundle.
@@ -153,98 +150,6 @@ class FormController extends FOSRestController implements ClassResourceInterface
     public function getBundleName()
     {
         return 'SuluFormBundle';
-    }
-
-    /**
-     * @param Request $request
-     *
-     * @return Response
-     */
-    public function cgetTemplateAction(Request $request)
-    {
-        $widths = [
-            [
-                'id' => 'full',
-                'name' => 'sulu_form.width.full',
-            ],
-            [
-                'id' => 'half',
-                'name' => 'sulu_form.width.half',
-            ],
-            [
-                'id' => 'one-third',
-                'name' => 'sulu_form.width.one-third',
-            ],
-            [
-                'id' => 'two-thirds',
-                'name' => 'sulu_form.width.two-thirds',
-            ],
-            [
-                'id' => 'one-quarter',
-                'name' => 'sulu_form.width.one-quarter',
-            ],
-            [
-                'id' => 'three-quarters',
-                'name' => 'sulu_form.width.three-quarters',
-            ],
-            [
-                'id' => 'one-sixth',
-                'name' => 'sulu_form.width.one-sixth',
-            ],
-            [
-                'id' => 'five-sixths',
-                'name' => 'sulu_form.width.five-sixths',
-            ],
-        ];
-
-        $types = $this->get('sulu_form.dynamic.form_field_type_pool')->all();
-        $receiverTypes = [
-            'to' => HelperInterface::MAIL_RECEIVER_TO,
-            'cc' => HelperInterface::MAIL_RECEIVER_CC,
-            'bcc' => HelperInterface::MAIL_RECEIVER_BCC,
-        ];
-
-        return $this->render(
-            $this->getBundleName() . ':' . $this->getListName() . ':template.html.twig',
-            [
-                'types' => $this->getSortedTypes($types),
-                'widths' => $widths,
-                'receiverTypes' => $receiverTypes,
-                'fallbackEmails' => [
-                    'from' => $this->getParameter('sulu_form.mail.from'),
-                    'to' => $this->getParameter('sulu_form.mail.to'),
-                ],
-            ]
-        );
-    }
-
-    /**
-     * @param FormFieldTypeInterface[] $types
-     *
-     * @return FormFieldTypeInterface[]
-     */
-    public function getSortedTypes($types = [])
-    {
-        /** @var Translator $translator */
-        $translator = $this->get('translator');
-        $locale = $this->getUser()->getLocale();
-
-        $sortedTypes = [];
-        $returnTypes = [];
-
-        $i = 0;
-        foreach ($types as $alias => $type) {
-            $translation = $translator->trans($type->getConfiguration()->getTitle(), [], 'backend', $locale);
-            $sortedTypes[$translation . $i] = ['alias' => $alias, 'type' => $type];
-            ++$i;
-        }
-
-        ksort($sortedTypes);
-        foreach ($sortedTypes as $sortedType) {
-            $returnTypes[$sortedType['alias']] = $sortedType['type'];
-        }
-
-        return $returnTypes;
     }
 
     /**
