+++ UPGRADE.md
@@ -1,5 +1,11 @@
 # Upgrade
 
+## 0.2.3
+
+### Form Template file moved
+
+`FormController:cgetTemplateAction` (/admin/api/forms/template) was moved to `TemplateController:formAction` (/admin/api//form/templates/form.html).
+
 ## 0.2.0
 
 ### Upgrade database schema
